RavenDB-12161 CPU usage file ordering
