+++ src/Raven.Server/Dashboard/MachineResourcesNotificationSender.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Raven.Server.Background;
 using Raven.Server.NotificationCenter;
-using Raven.Server.Utils;
+using Raven.Server.Utils.Cpu;
 using Sparrow;
 using Sparrow.Collections;
 using Sparrow.LowMemory;
