+++ test/SlowTests/Issues/RavenDB_12161.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Runtime.InteropServices;
 using FastTests;
-using Raven.Server.Utils;
+using Raven.Server.Utils.Cpu;
 using Sparrow.Json;
 using Sparrow.Logging;
 using Xunit;
